Factor out JSON logic from KubernetesClient
- Add circe-yaml to prep for sending rendered blueprints to schedulers
- Stub KubernetesHttpV2 for blueprint-aware Kubernetes interpreter
- Bump snakeyaml dependency since circe-yaml wants a slightly newer version
    - Tiny API-compatibility change in YamlParser
