+++ core/build.sbt
@@ -33,7 +33,7 @@ libraryDependencies ++= Seq(
   "io.verizon.knobs"           %% "core"                               % V.knobs,
   "io.verizon.journal"         %% "core"                               % V.journal,
   "io.verizon.quiver"          %% "core"                               % "7.0.19",
-  "org.yaml"                    % "snakeyaml"                          % "1.16",
+  "org.yaml"                    % "snakeyaml"                          % "1.20",
   "org.scalatra.scalate"       %% "scalate-core"                       % "1.9.0",
   "io.argonaut"                %% "argonaut"                           % "6.2.1",
   "io.argonaut"                %% "argonaut-cats"                      % "6.2.1",
@@ -44,6 +44,7 @@ libraryDependencies ++= Seq(
   "org.tpolecat"               %% "doobie-hikari"                      % V.doobie,
   "org.spire-math"             %% "spire"                              % "0.11.0",
   "org.flywaydb"                % "flyway-core"                        % "3.2.1",
+  "io.circe"                   %% "circe-yaml"                         % "0.8.0",
   "net.databinder.dispatch"    %% "dispatch-core"                      % "0.11.2",
   "ca.mrvisser"                %% "sealerate"                          % "0.0.4",
   "org.typelevel"              %% "cats-kernel-laws"                   % V.cats   % "test",
