+++ core/src/main/scala/KubernetesClient.scala
@@ -1,9 +1,6 @@
 package nelson
 
-import nelson.Datacenter.StackName
 import nelson.Infrastructure.KubernetesMode
-import nelson.Manifest.{HealthCheck => HealthProbe, _}
-import nelson.docker.Docker.Image
 import nelson.health._
 
 import argonaut._
@@ -58,28 +55,24 @@ final class KubernetesClient(version: KubernetesVersion, endpoint: Uri, client: 
     case KubernetesMode.OutCluster(_, token) => token
   }
 
-  def createDeployment(namespace: String, stackName: StackName, image: Image, plan: Plan, ports: Option[Ports]): IO[Json] = {
-    val json = KubernetesJson.deployment(namespace,stackName, image, plan, ports)
+  def createDeployment(namespace: String, payload: Json): IO[Json] = {
     val request = addCreds(Request(Method.POST, deploymentUri(namespace)))
-    client.expect[Json](request.withBody(json))
+    client.expect[Json](request.withBody(payload))
   }
 
-  def createService(namespace: String, stackName: StackName, ports: Option[Ports]): IO[Json] = {
-    val json = KubernetesJson.service(namespace, stackName, ports)
+  def createService(namespace: String, payload: Json): IO[Json] = {
     val request = addCreds(Request(Method.POST, serviceUri(namespace)))
-    client.expect[Json](request.withBody(json))
+    client.expect[Json](request.withBody(payload))
   }
 
-  def createCronJob(namespace: String, stackName: StackName, image: Image, plan: Plan, cronExpr: String): IO[Json] = {
-    val json = KubernetesJson.cronJob(namespace, stackName, image, plan, cronExpr)
+  def createCronJob(namespace: String, payload: Json): IO[Json] = {
     val request = addCreds(Request(Method.POST, cronJobUri(namespace)))
-    client.expect[Json](request.withBody(json))
+    client.expect[Json](request.withBody(payload))
   }
 
-  def createJob(namespace: String, stackName: StackName, image: Image, plan: Plan): IO[Json] = {
-    val json = KubernetesJson.job(namespace, stackName, image, plan)
+  def createJob(namespace: String, payload: Json): IO[Json] = {
     val request = addCreds(Request(Method.POST, jobUri(namespace)))
-    client.expect[Json](request.withBody(json))
+    client.expect[Json](request.withBody(payload))
   }
 
   def deleteDeployment(namespace: String, name: String): IO[Json] = {
@@ -183,229 +176,6 @@ object KubernetesClient {
 }
 
 object KubernetesJson {
-  val defaultCPU = 0.5
-  val defaultMemory = 512
-
-  def resources(cpu: ResourceSpec, memory: ResourceSpec): JsonObject = {
-    val (cpuRequest, cpuLimit) = cpu.fold(
-      (defaultCPU, defaultCPU),
-      limit => (limit, limit),
-      (request, limit) => (request, limit)
-    )
-    val (memoryRequest, memoryLimit) = memory.fold(
-      (s"${defaultMemory}M", s"${defaultMemory}M"),
-      limit => (s"${limit}M", s"${limit}M"),
-      (request, limit) => (s"${request}M", s"${limit}M")
-    )
-
-    JsonObject.single("resources", argonaut.Json(
-      "requests" := argonaut.Json(
-        "cpu"    := cpuRequest,
-        "memory" := memoryRequest
-      ),
-      "limits" := argonaut.Json(
-        "cpu"    := cpuLimit,
-        "memory" := memoryLimit
-      )
-    ))
-  }
-
-  def volumeJson(volume: Volume): Json = argonaut.Json.jObject(volume match {
-    case Volume(id, _, sizeInMb) =>
-      JsonObject.fromTraversableOnce(List(
-        "name" := id,
-        "emptyDir" := argonaut.Json("sizeLimit" := s"${sizeInMb}M")
-      ))
-  })
-
-  def volumesJson(volumes: List[Volume]): JsonObject = volumes match {
-    case Nil     => JsonObject.empty
-    case v :: vs => JsonObject.single("volumes", (volumeJson(v) :: vs.map(volumeJson)).asJson)
-  }
-
-  def volumeMountJson(volume: Volume): Json = argonaut.Json.jObject(volume match {
-    case Volume(id, mountPath, _) =>
-      JsonObject.fromTraversableOnce(List(
-        "name" := id,
-        "mountPath" := mountPath.toString
-      ))
-  })
-
-  def volumeMountsJson(volumes: List[Volume]): JsonObject = volumes match {
-    case Nil => JsonObject.empty
-    case v :: vs => JsonObject.single("volumeMounts", (volumeMountJson(v) :: vs.map(volumeMountJson)).asJson)
-  }
-
-  def deployment(
-    namespace: String,
-    stackName: StackName,
-    image:     Image,
-    plan:      Plan,
-    ports:     Option[Ports]
-  ): Json =
-    argonaut.Json(
-      "kind"       := "Deployment",
-      "metadata"   := argonaut.Json(
-        "name"      := stackName.toString,
-        "namespace" := namespace,
-        "labels"    := argonaut.Json(
-          "stackName"   := stackName.toString,
-          "serviceName" := stackName.serviceType,
-          "version"     := stackName.version.toString,
-          "nelson"      := "true"
-        )
-      ),
-      "spec" := argonaut.Json(
-        "replicas" := plan.environment.desiredInstances.getOrElse(1),
-        "selector" := argonaut.Json(
-          "matchLabels" := argonaut.Json("stackName" := stackName.toString)
-        ),
-        "template" := argonaut.Json(
-          "metadata" := argonaut.Json(
-            "labels" := argonaut.Json(
-              "stackName"   := stackName.toString,
-              "serviceName" := stackName.serviceType,
-              "version"     := stackName.version.toString,
-              "nelson"      := "true"
-            )
-          ),
-          "spec" := argonaut.Json.jObject(combineJsonObjects(
-            argonaut.JsonObject.fromTraversableOnce(List("containers" := List(
-              argonaut.Json.jObject(combineJsonObjects(JsonObject.fromTraversableOnce(List(
-                "name"      := stackName.toString,
-                "image"     := image.toString,
-                "env"       := plan.environment.bindings,
-                "ports"     := containerPortsJson(ports.toList.flatMap(_.nel.toList))
-              )), volumeMountsJson(plan.environment.volumes), resources(plan.environment.cpu, plan.environment.memory), livenessProbe(plan.environment.healthChecks)))
-            ))),
-            volumesJson(plan.environment.volumes)
-          ))
-        )
-      )
-    )
-
-  def service(namespace: String, stackName: StackName, ports: Option[Ports]): Json =
-    argonaut.Json(
-      "kind"       := "Service",
-      "metadata"   := argonaut.Json(
-        "name"      := stackName.toString,
-        "namespace" := namespace,
-        "labels"    := argonaut.Json(
-          "stackName"   := stackName.toString,
-          "serviceName" := stackName.serviceType,
-          "version"     := stackName.version.toString,
-          "nelson"      := "true"
-        )
-      ),
-      "spec" := argonaut.Json(
-        "selector" := argonaut.Json("stackName" := stackName.toString),
-        "ports"    := servicePortsJson(ports.toList.flatMap(_.nel.toList)),
-        "type"     := "ClusterIP"
-      )
-    )
-
-  def cronJob(
-    namespace: String,
-    stackName: StackName,
-    image:     Image,
-    plan:      Plan,
-    cronExpr:  String
-  ): Json =
-    argonaut.Json(
-      "kind"       := "CronJob",
-      "metadata"   := argonaut.Json(
-        "name"      := stackName.toString,
-        "namespace" := namespace,
-        "labels"    := argonaut.Json(
-          "stackName"   := stackName.toString,
-          "serviceName" := stackName.serviceType,
-          "version"     := stackName.version.toString,
-          "nelson"      := "true"
-        )
-      ),
-      "spec" := argonaut.Json(
-        "schedule"    := cronExpr,
-        "jobTemplate" := argonaut.Json.jObject(jobSpecJson(stackName, image, plan))
-      )
-    )
-
-  def job(
-    namespace: String,
-    stackName: StackName,
-    image:     Image,
-    plan:      Plan
-  ): Json =
-    argonaut.Json.jObject(combineJsonObjects(JsonObject.fromTraversableOnce(List(
-      "kind"       := "Job",
-      "metadata"   := argonaut.Json(
-        "name"      := stackName.toString,
-        "namespace" := namespace,
-        "labels"    := argonaut.Json(
-          "stackName"   := stackName.toString,
-          "serviceName" := stackName.serviceType,
-          "version"     := stackName.version.toString,
-          "nelson"      := "true"
-        )
-      )
-    )), jobSpecJson(stackName, image, plan)))
-
-  private def jobSpecJson(
-    stackName: StackName,
-    image:     Image,
-    plan:      Plan
-  ): JsonObject = {
-    val backoffLimit = JsonObject.single("backoffLimit", plan.environment.retries.getOrElse(3).asJson)
-
-    JsonObject.fromTraversableOnce(List(
-      "spec" := argonaut.Json.jObject(combineJsonObjects(JsonObject.fromTraversableOnce(List(
-        "completions"  := plan.environment.desiredInstances.getOrElse(1),
-        "template"     := argonaut.Json(
-          "metadata" := argonaut.Json(
-            "name" := stackName.toString,
-            "labels" := argonaut.Json(
-              "stackName"   := stackName.toString,
-              "serviceName" := stackName.serviceType,
-              "version"     := stackName.version.toString,
-              "nelson"      := "true"
-            )
-          ),
-          "spec" := argonaut.Json.jObject(combineJsonObjects(
-            JsonObject.fromTraversableOnce(List("containers" := List(
-              argonaut.Json.jObject(combineJsonObjects(JsonObject.fromTraversableOnce(List(
-                "name"  := stackName.toString,
-                "image" := image.toString,
-                "env"   := plan.environment.bindings
-              )), volumeMountsJson(plan.environment.volumes), resources(plan.environment.cpu, plan.environment.memory)))
-            ))),
-            volumesJson(plan.environment.volumes)) :+
-            // See: https://kubernetes.io/docs/concepts/workloads/controllers/jobs-run-to-completion/#pod-backoff-failure-policy
-            // This should be "OnFailure" but at the time of this writing this section said:
-            // Note: Due to a known issue #54870, when the spec.template.spec.restartPolicy field is set to “OnFailure”,
-            // the back-off limit may be ineffective. As a short-term workaround, set the restart policy for the embedded template to “Never”
-            // https://github.com/kubernetes/kubernetes/issues/54870
-            ("restartPolicy" := "Never")
-          )
-        )
-      )), backoffLimit))
-    ))
-  }
-
-  // HealthChecks in Nelson seem to correspond to Kubernetes liveness probes
-  // Kubernetes seems to only support one one liveness probe, so take the first one..
-  private def livenessProbe(healthChecks: List[HealthProbe]): JsonObject =
-    healthChecks.headOption match {
-      case None => JsonObject.empty
-      case Some(HealthProbe(_, portRef, _, path, interval, timeout)) =>
-        JsonObject.single("livenessProbe", argonaut.Json(
-          "httpGet" := argonaut.Json(
-            "path" := path.getOrElse("/"),
-            "port" := portRef
-          ),
-          "periodSeconds"  := interval.toSeconds,
-          "timeoutSeconds" := timeout.toSeconds
-        ))
-    }
-
   // Status seems to be largely undocumented in the K8s docs, so your best bet is to stare at
   // https://github.com/kubernetes/kubernetes/tree/master/api/openapi-spec
   // Recommend using 'jq' for your sanity
@@ -458,23 +228,6 @@ object KubernetesJson {
   private def combineJsonObjects(x: JsonObject, xs: JsonObject*): JsonObject =
     JsonObject.fromTraversableOnce(x.toList ++ xs.toList.flatMap(_.toList))
 
-  private def containerPortsJson(ports: List[Port]): List[Json] =
-    ports.map { port =>
-      argonaut.Json(
-        "name"          := port.ref,
-        "containerPort" := port.port
-      )
-    }
-
-  private def servicePortsJson(ports: List[Port]): List[Json] =
-    ports.map { port =>
-      argonaut.Json(
-        "name"       := port.ref,
-        "port"       := port.port,
-        "targetPort" := port.port
-      )
-    }
-
   val healthStatusDecoder: DecodeJson[HealthStatus] =
     DecodeJson(c =>
       for {
@@ -494,11 +247,4 @@ object KubernetesJson {
     case "False" => Failing
     case _       => Unknown
   }
-
-  private implicit val envVarEncoder: EncodeJson[EnvironmentVariable] =
-    EncodeJson { (ev: EnvironmentVariable) =>
-      ("name"  := ev.name)  ->:
-      ("value" := ev.value) ->:
-      jEmptyObject
-    }
 }
